add run_process and remove modulestorage
