+++ src/prefect/cli/__init__.py
@@ -10,7 +10,7 @@ import prefect.cli.artifact
 import prefect.cli.block
 import prefect.cli.cloud
 import prefect.cli.cloud.webhook
-import prefect.cli.command
+import prefect.cli.shell
 import prefect.cli.concurrency_limit
 import prefect.cli.config
 import prefect.cli.deploy
